fragments, scrapFragments を分割
